+++ scripts/prepare_perfect_foresight.py
@@ -18,6 +18,7 @@ from _helpers import (
     update_config_from_wildcards,
 )
 from add_existing_baseyear import add_build_year_to_new_assets
+from prepare_sector_network import adjust_transport_temporal_agg
 from pypsa.descriptors import expand_series
 from pypsa.io import import_components_from_dataframe
 from six import iterkeys
@@ -165,7 +166,8 @@ def concat_networks(years):
         year = years[i]
         network = pypsa.Network(network_path)
         adjust_electricity_grid(network, year, years)
-        add_build_year_to_new_assets(network, year)
+        if not i == 0:
+            add_build_year_to_new_assets(network, year)
 
         # static ----------------------------------
         for component in network.iterate_components(
@@ -520,6 +522,7 @@ if __name__ == "__main__":
     segments = snakemake.params.time_resolution
     if isinstance(segments, (int, float)):
         n = apply_time_segmentation_perfect(n, segments, solver_name=solver_name)
+        adjust_transport_temporal_agg(n)
 
     # adjust global constraints lv limit if the same for all years
     n = adjust_lvlimit(n)